fixing compiler warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@5263 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/World.h
@@ -358,7 +358,7 @@ inline Flag&		World::getFlag(int index) const
 inline const float*	World::getBase(int team, int base) const
 {
   const TeamBases &b = bases[team];
-  if ((base < 0) || (base >= b.size()))
+  if ((base < 0) || (base >= (int)b.size()))
     return NULL;
 
   return b[base].p;
